+++ TMS.Infrastructure/Persistence/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -21,9 +21,9 @@ namespace TMS.Infrastructure.Persistence.Migrations
 
             modelBuilder.Entity("TMS.Domain.Entities.Employee", b =>
                 {
-                    b.Property<int>("EmployeeId")
+                    b.Property<long>("EmployeeId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<bool>("Active")
@@ -58,13 +58,13 @@ namespace TMS.Infrastructure.Persistence.Migrations
 
             modelBuilder.Entity("TMS.Domain.Entities.Issue", b =>
                 {
-                    b.Property<int>("IssueId")
+                    b.Property<long>("IssueId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int?>("AssigneeId")
-                        .HasColumnType("int");
+                    b.Property<long?>("AssigneeId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(1024)")
@@ -78,8 +78,8 @@ namespace TMS.Infrastructure.Persistence.Migrations
                     b.Property<int>("Priority")
                         .HasColumnType("int");
 
-                    b.Property<int?>("ReporterId")
-                        .HasColumnType("int");
+                    b.Property<long?>("ReporterId")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("Status")
                         .HasColumnType("int");
